refactor: extract transfer error classes and recovery checks into transfer-errors module
